+++ public/constants.ts
@@ -10,15 +10,11 @@ export const MESSAGE_CHANNELS = {
     AUTHENTICATION_GET_PROFILE_TOKEN: 'authentication_get_profile_token',
     AUTHENTICATION_LOGIN: 'authentication_login',
     AUTHENTICATION_LOGOUT: 'authentication_logout',
-    DIRECTORY_GET_DIRECTORIES: 'directory_getDirectories',
-    MODEL_REPOSITORY_GET_DEFINITION: 'model_definition',
     SETTING_HIGH_CONTRAST: 'setting_highContrast',
 };
 
 export const API_INTERFACES = {
     AUTHENTICATION: 'api_authentication',
     DEVICE: 'api_device',
-    DIRECTORY: 'api_directory',
-    MODEL_DEFINITION: 'api_modelDefinition',
     SETTINGS: 'api_settings'
 };
