--- public/interfaces/directoryInterface.ts
@@ -1,10 +1,0 @@
-/***********************************************************
- * Copyright (c) Microsoft Corporation. All rights reserved.
- * Licensed under the MIT License
- **********************************************************/
-export interface GetDirectoriesParameters {
-    path: string;
-}
-export interface DirectoryInterface {
-    getDirectories(params: GetDirectoriesParameters): Promise<string[]>;
-}
