+++ src/app/api/shared/interfaceUtils.ts
@@ -3,13 +3,10 @@
  * Licensed under the MIT License
  **********************************************************/
 import { SettingsInterface } from '../../../../public/interfaces/settingsInterface';
-import { DirectoryInterface } from '../../../../public/interfaces/directoryInterface';
-import { ModelRepositoryInterface } from '../../../../public/interfaces/modelRepositoryInterface';
 import { AuthenticationInterface } from './../../../../public/interfaces/authenticationInterface';
 import { API_INTERFACES } from '../../../../public/constants';
 import { appConfig, HostMode } from '../../../appConfig/appConfig';
 import { HIGH_CONTRAST } from '../../constants/browserStorage';
-import { LocalRepoServiceHandler } from '../handlers/localRepoServiceHandler';
 import { PublicDigitalTwinsModelRepoHelper, PublicDigitalTwinsModelInterface } from '../services/publicDigitalTwinsModelRepoHelper';
 
 export const NOT_AVAILABLE = 'Feature is not available in this configuration';
@@ -27,22 +24,6 @@ export const getSettingsInterfaceForBrowser = (): SettingsInterface => {
             return Promise.resolve(result);
         }
     });
-};
-
-export const getLocalModelRepositoryInterface = (): ModelRepositoryInterface => {
-    if (appConfig.hostMode !== HostMode.Electron) {
-        return new LocalRepoServiceHandler();
-    }
-
-    return getElectronInterface(API_INTERFACES.MODEL_DEFINITION);
-};
-
-export const getDirectoryInterface = (): DirectoryInterface => {
-    if (appConfig.hostMode !== HostMode.Electron) {
-        return new LocalRepoServiceHandler();
-    }
-
-    return getElectronInterface(API_INTERFACES.DIRECTORY);
 };
 
 export const getPublicDigitalTwinsModelInterface = (): PublicDigitalTwinsModelInterface => {
