+++ src/app/constants/repositoryLocationTypes.ts
@@ -6,5 +6,5 @@ export enum REPOSITORY_LOCATION_TYPE {
     Configurable = 'CONFIGURABLE',
     Public = 'PUBLIC',
     Local = 'LOCAL',
-    LocalDMR = 'LOCAL_DMR'
+    LocalDMR = 'DMR'
 }
