+++ src/app/modelRepository/components/modelRepositoryInstruction.tsx
@@ -10,12 +10,20 @@ import { ResourceKeys } from '../../../localization/resourceKeys';
 import { ROUTE_PARTS } from '../../constants/routes';
 import './modelRepositoryInstruction.scss';
 
-export interface ModelRepositoryInstructionDataProps {
-    empty: boolean;
-}
-
-export const ModelRepositoryInstruction: React.FC<ModelRepositoryInstructionDataProps> = props => {
+export const ModelRepositoryInstruction: React.FC = () => {
     const { t } = useTranslation();
+
+    const privacyStatement =  (
+        <div className="privacy-statement">
+            <span>{t(ResourceKeys.settings.questions.questions.privacy.text)}</span>
+            <Link
+                href={t(ResourceKeys.settings.questions.questions.privacy.link)}
+                target="_blank"
+            >
+                {t(ResourceKeys.settings.questions.questions.privacy.linkText)}
+            </Link>
+        </div>
+    );
 
     return (
         <div className="model-repository-instruction">
@@ -31,23 +39,10 @@ export const ModelRepositoryInstruction: React.FC<ModelRepositoryInstructionData
             >
                 {t(ResourceKeys.modelRepository.description.help)}
             </Link>
+
             <h3 role="heading" aria-level={1}>{t(ResourceKeys.modelRepository.description.header)}</h3>
             {t(ResourceKeys.modelRepository.instruction)}
-            {RenderPrivaryStatement()}
-        </div>);
-};
-
-const RenderPrivaryStatement = () => {
-    const { t } = useTranslation();
-    return (
-        <div className="privacy-statement">
-            <span>{t(ResourceKeys.settings.questions.questions.privacy.text)}</span>
-            <Link
-                href={t(ResourceKeys.settings.questions.questions.privacy.link)}
-                target="_blank"
-            >
-               {t(ResourceKeys.settings.questions.questions.privacy.linkText)}
-            </Link>
+            {privacyStatement}
         </div>
     );
 };
