+++ src/app/modelRepository/components/modelRepositoryLocationList.tsx
@@ -5,44 +5,24 @@
 import * as React from 'react';
 import { Container, Draggable } from 'react-smooth-dnd';
 import { ModelRepositoryLocationListItem } from './modelRepositoryLocationListItem';
-import { ModelRepositoryConfiguration } from '../../shared/modelRepository/state';
-import { StringMap } from '../../api/models/stringMap';
+import { ModelRepositoryConfiguration, ModelRepositoryStateInterface } from '../../shared/modelRepository/state';
+import { ModelRepositoryFormType } from '../hooks/useModelRepositoryForm';
+import { validateRepositoryLocationSettings } from './commands';
 import './modelRepositoryLocationList.scss';
 
-export interface ModelRepositoryLocationListProps {
-    repositoryLocationSettings: ModelRepositoryConfiguration[];
-    repositoryLocationSettingsErrors: StringMap<string>;
-    onChangeRepositoryLocationSettings(settings: ModelRepositoryConfiguration[]): void;
-}
-
-export const ModelRepositoryLocationList: React.FC<ModelRepositoryLocationListProps> = props => {
-    const { repositoryLocationSettings, repositoryLocationSettingsErrors, onChangeRepositoryLocationSettings} = props;
+export const ModelRepositoryLocationList: React.FC<{formState: ModelRepositoryFormType}> = ({formState}) => {
+    const[{repositoryLocationSettings }, {setRepositoryLocationSettings, setRepositoryLocationSettingsErrors, setDirtyFlag}] = formState;
+    const onChangeRepositoryLocationSettings = (updatedRepositoryLocationSettings: ModelRepositoryStateInterface) => {
+        setDirtyFlag(true);
+        setRepositoryLocationSettingsErrors(validateRepositoryLocationSettings(updatedRepositoryLocationSettings));
+        setRepositoryLocationSettings([
+            ...updatedRepositoryLocationSettings
+        ]);
+    };
 
     const onDrop = (e: {addedIndex: number, removedIndex: number}) => {
         const updatedRepositoryLocationSettings = [...repositoryLocationSettings];
         updatedRepositoryLocationSettings.splice(e.addedIndex, 0, updatedRepositoryLocationSettings.splice(e.removedIndex, 1)[0]);
-
-        onChangeRepositoryLocationSettings(updatedRepositoryLocationSettings);
-    };
-
-    const onRemoveRepositoryLocationSetting = (index: number) => {
-        const updatedRepositoryLocationSettings = [...props.repositoryLocationSettings];
-        updatedRepositoryLocationSettings.splice(index, 1);
-
-        onChangeRepositoryLocationSettings(updatedRepositoryLocationSettings);
-    };
-
-    const onChangeRepositoryLocationSettingValue = (index: number, value: string) => {
-        const updatedRepositoryLocationSettings = repositoryLocationSettings.map((setting, i) => {
-            if (i === index) {
-                const updatedSetting = {...setting};
-                updatedSetting.value = value;
-                return updatedSetting;
-            } else {
-                return setting;
-            }
-        });
-
         onChangeRepositoryLocationSettings(updatedRepositoryLocationSettings);
     };
 
@@ -52,9 +32,7 @@ export const ModelRepositoryLocationList: React.FC<ModelRepositoryLocationListPr
                     <ModelRepositoryLocationListItem
                         index={index}
                         item={item}
-                        errorKey={repositoryLocationSettingsErrors[item.repositoryLocationType]}
-                        onChangeRepositoryLocationSettingValue={onChangeRepositoryLocationSettingValue}
-                        onRemoveRepositoryLocationSetting={onRemoveRepositoryLocationSetting}
+                        formState={formState}
                     />
                 </Draggable>
         );
