+++ src/app/modelRepository/components/modelRepositoryLocationListItem.scss
@@ -17,9 +17,6 @@
     .ms-Dialog-header{
         word-wrap: break-word;
     }
-    .labelSection {
-        text-overflow: ellipsis;
-    }
     .folder-button {
         width: 100%;
         text-align: left;
