+++ src/app/modelRepository/components/modelRepositoryLocationListItem.tsx
@@ -4,241 +4,64 @@
  **********************************************************/
 import * as React from 'react';
 import { useTranslation } from 'react-i18next';
-import { IconButton, DefaultButton, PrimaryButton, Label, Dialog, DialogFooter, TextField } from '@fluentui/react';
+import { IconButton } from '@fluentui/react';
 import { REPOSITORY_LOCATION_TYPE } from '../../constants/repositoryLocationTypes';
 import { ResourceKeys } from '../../../localization/resourceKeys';
-import { CANCEL, NAVIGATE_BACK, FOLDER } from '../../constants/iconNames';
-import { fetchDirectories } from '../../api/services/localRepoService';
-import { LabelWithRichCallout } from '../../shared/components/labelWithRichCallout';
-import { getRootFolder, getParentFolder } from '../../shared/utils/utils';
-import { ModelRepositoryConfiguration } from '../../shared/modelRepository/state';
-import { PUBLIC_REPO_HOSTNAME } from '../../constants/apiConstants';
+import { CANCEL } from '../../constants/iconNames';
+import { ModelRepositoryConfiguration, ModelRepositoryStateInterface } from '../../shared/modelRepository/state';
+import { ModelRepositoryFormType } from '../hooks/useModelRepositoryForm';
+import { validateRepositoryLocationSettings } from './commands';
+import { ListItemPublicRepo } from './listItemPublicRepo';
+import { ListItemConfigurableRepo } from './listItemConfigurableRepo';
+import { ListItemLocal } from './listItemLocal';
 import './modelRepositoryLocationListItem.scss';
 
 export interface ModelRepositoryLocationListItemProps {
-    errorKey?: string;
     index: number;
     item: ModelRepositoryConfiguration;
-    onChangeRepositoryLocationSettingValue: (index: number, path: string) => void;
-    onRemoveRepositoryLocationSetting: (index: number) => void;
+    formState: ModelRepositoryFormType;
 }
 
-export interface RepositoryLocationListItemState {
-    currentFolder: string;
-    subFolders: string[];
-    showFolderPicker: boolean;
-    showError: boolean;
-}
-
-// tslint:disable-next-line: cyclomatic-complexity
-export const ModelRepositoryLocationListItem: React.FC<ModelRepositoryLocationListItemProps> = (props: ModelRepositoryLocationListItemProps) => {
+export const ModelRepositoryLocationListItem: React.FC<ModelRepositoryLocationListItemProps> = ({index, item, formState}) => {
     const { t } = useTranslation();
+    const [{repositoryLocationSettings }, {setRepositoryLocationSettings, setRepositoryLocationSettingsErrors, setDirtyFlag}] = formState;
 
-    const { item, index, onChangeRepositoryLocationSettingValue, onRemoveRepositoryLocationSetting } = props;
-    let initialCurrentFolder = '';
-    if (item && item.repositoryLocationType === REPOSITORY_LOCATION_TYPE.Local) {
-        initialCurrentFolder = item.value || getRootFolder();
-    }
-    let initialConfigurableRepositoryPath = '';
-    if (item && item.repositoryLocationType === REPOSITORY_LOCATION_TYPE.Configurable) {
-        initialConfigurableRepositoryPath = item.value;
-    }
+    const onRemove = () => {
+        const updatedRepositoryLocationSettings = [...repositoryLocationSettings];
+        updatedRepositoryLocationSettings.splice(index, 1);
+        onChangeRepositoryLocationSettings(updatedRepositoryLocationSettings);
+    };
 
-    const [ subFolders, setSubFolders ] = React.useState([]);
-    const [ currentFolder, setCurrentFolder ] = React.useState(initialCurrentFolder);
-    const [ currentConfigurableRepositoryPath, setCurrentConfigurableRepositoryPath ] = React.useState(initialConfigurableRepositoryPath);
-    const [ showError, setShowError ] = React.useState<boolean>(false);
-    const [ showFolderPicker, setShowFolderPicker ] = React.useState<boolean>(false);
-
-    React.useEffect(() => {
-        setCurrentConfigurableRepositoryPath(initialConfigurableRepositoryPath);
-    }, [initialConfigurableRepositoryPath]); // tslint:disable-line: align
-
-    React.useEffect(() => {
-        setCurrentFolder(initialCurrentFolder);
-    }, [initialCurrentFolder]); // tslint:disable-line: align
-
-    const onRemove = () => onRemoveRepositoryLocationSetting(index);
+    const onChangeRepositoryLocationSettings = (updatedRepositoryLocationSettings: ModelRepositoryStateInterface) => {
+        setDirtyFlag(true);
+        setRepositoryLocationSettingsErrors(validateRepositoryLocationSettings(updatedRepositoryLocationSettings));
+        setRepositoryLocationSettings([
+            ...updatedRepositoryLocationSettings
+        ]);
+    };
 
     const renderItemDetail = () => {
         return (
             <div className="item-details">
                 {item.repositoryLocationType === REPOSITORY_LOCATION_TYPE.Public &&
-                    renderPublicRepoItem()
+                    <ListItemPublicRepo/>
                 }
-                {item.repositoryLocationType === REPOSITORY_LOCATION_TYPE.Local &&
-                    renderLocalFolderItem()
+                {(item.repositoryLocationType === REPOSITORY_LOCATION_TYPE.Local || item.repositoryLocationType === REPOSITORY_LOCATION_TYPE.LocalDMR) &&
+                    <ListItemLocal
+                        index={index}
+                        item={item}
+                        formState={formState}
+                        repoType={item.repositoryLocationType}
+                    />
                 }
                 {item.repositoryLocationType === REPOSITORY_LOCATION_TYPE.Configurable &&
-                    renderConfigurableRepoItem()
+                    <ListItemConfigurableRepo
+                        index={index}
+                        item={item}
+                        formState={formState}
+                    />
                 }
             </div>);
-    };
-
-    const renderPublicRepoItem = () => {
-        return(
-            <>
-                <div className="labelSection">
-                    <Label>{t(ResourceKeys.modelRepository.types.public.label)}</Label>
-                </div>
-                <TextField
-                    className="local-folder-textbox"
-                    label={t(ResourceKeys.modelRepository.types.configurable.textBoxLabel)}
-                    ariaLabel={t(ResourceKeys.modelRepository.types.configurable.textBoxLabel)}
-                    value={PUBLIC_REPO_HOSTNAME}
-                    readOnly={true}
-                    prefix="https://"
-                />
-            </>
-        );
-    };
-
-    const renderLocalFolderItem = () => {
-        return(
-                <>
-                    <div className="labelSection">
-                        <LabelWithRichCallout
-                            calloutContent={t(ResourceKeys.modelRepository.types.local.infoText)}
-                            required={true}
-                        >
-                            {t(ResourceKeys.modelRepository.types.local.label)}
-                        </LabelWithRichCallout>
-                    </div>
-                    <TextField
-                        className="local-folder-textbox"
-                        label={t(ResourceKeys.modelRepository.types.local.textBoxLabel)}
-                        ariaLabel={t(ResourceKeys.modelRepository.types.local.textBoxLabel)}
-                        value={currentFolder}
-                        readOnly={false}
-                        errorMessage={props.errorKey ? t(props.errorKey) : ''}
-                        onChange={onFolderPathChange}
-                    />
-                    <DefaultButton
-                        className="local-folder-launch"
-                        text={t(ResourceKeys.modelRepository.types.local.folderPicker.command.openPicker)}
-                        ariaLabel={t(ResourceKeys.modelRepository.types.local.folderPicker.command.openPicker)}
-                        onClick={onShowFolderPicker}
-                    />
-                    {renderFolderPicker()}
-                </>
-            );
-    };
-
-    const renderConfigurableRepoItem = () => {
-        return(
-            <>
-                <div className="labelSection">
-                    <LabelWithRichCallout
-                        calloutContent={t(ResourceKeys.modelRepository.types.configurable.infoText)}
-                        required={true}
-                    >
-                        {t(ResourceKeys.modelRepository.types.configurable.label)}
-                    </LabelWithRichCallout>
-                </div>
-                <TextField
-                    className="local-folder-textbox"
-                    label={t(ResourceKeys.modelRepository.types.configurable.textBoxLabel)}
-                    ariaLabel={t(ResourceKeys.modelRepository.types.configurable.textBoxLabel)}
-                    value={currentConfigurableRepositoryPath}
-                    readOnly={false}
-                    errorMessage={props.errorKey ? t(props.errorKey) : ''}
-                    onChange={repositoryEndpointChange}
-                    prefix="https://"
-                />
-            </>
-        );
-    };
-
-    const repositoryEndpointChange = (event: React.FormEvent<HTMLInputElement | HTMLTextAreaElement>, newValue?: string) => {
-        setCurrentConfigurableRepositoryPath(newValue);
-        onChangeRepositoryLocationSettingValue(index, newValue);
-   };
-
-    const onFolderPathChange = (event: React.FormEvent<HTMLInputElement | HTMLTextAreaElement>, newValue?: string) => {
-        setCurrentFolder(newValue);
-        onChangeRepositoryLocationSettingValue(index, newValue);
-   };
-
-    const onShowFolderPicker = () => {
-        fetchSubDirectoriesInfo(currentFolder);
-        setShowFolderPicker(true);
-    };
-
-    const dismissFolderPicker = () => {
-        setCurrentFolder(item.value || getRootFolder());
-        setShowFolderPicker(false);
-    };
-
-    const onSelectFolder = () => {
-        onChangeRepositoryLocationSettingValue(index, currentFolder);
-        setShowFolderPicker(false);
-    };
-
-    const onClickFolderName = (folder: string) => () => {
-        const newDir = currentFolder ? `${currentFolder.replace(/\/$/, '')}/${folder}` : folder;
-        setCurrentFolder(newDir);
-        fetchSubDirectoriesInfo(newDir);
-    };
-
-    const onNavigateBack = () => {
-        const parentFolder = getParentFolder(currentFolder);
-        setCurrentFolder(parentFolder);
-        fetchSubDirectoriesInfo(parentFolder);
-    };
-
-    const fetchSubDirectoriesInfo = (folderName: string) => {
-        fetchDirectories(folderName).then(result => {
-            setShowError(false);
-            setSubFolders(result);
-        }).catch(error => {
-            setShowError(true);
-        });
-    };
-
-    const renderFolderPicker = () => {
-        return (
-            <div role="dialog">
-                <Dialog
-                    hidden={!showFolderPicker}
-                    title={t(ResourceKeys.modelRepository.types.local.folderPicker.dialog.title)}
-                    modalProps={{
-                        className: 'folder-picker-dialog',
-                        isBlocking: false
-                    }}
-                    dialogContentProps={{
-                        subText: currentFolder && t(ResourceKeys.modelRepository.types.local.folderPicker.dialog.subText, {folder: currentFolder})
-                    }}
-                    onDismiss={dismissFolderPicker}
-                >
-                    <div className="folder-links">
-                        <DefaultButton
-                            className="folder-button"
-                            iconProps={{ iconName: NAVIGATE_BACK }}
-                            text={t(ResourceKeys.modelRepository.types.local.folderPicker.command.navigateToParent)}
-                            ariaLabel={t(ResourceKeys.modelRepository.types.local.folderPicker.command.navigateToParent)}
-                            onClick={onNavigateBack}
-                            disabled={currentFolder === getRootFolder()}
-                        />
-                        {showError ? <div className="no-folders-text">{t(ResourceKeys.modelRepository.types.local.folderPicker.dialog.error)}</div> :
-                            subFolders && subFolders.length > 0 ?
-                                subFolders.map(folder =>
-                                    <DefaultButton
-                                        className="folder-button"
-                                        iconProps={{ iconName: FOLDER }}
-                                        key={folder}
-                                        text={folder}
-                                        onClick={onClickFolderName(folder)}
-                                    />)
-                                :
-                                <div className="no-folders-text">{t(ResourceKeys.modelRepository.types.local.folderPicker.dialog.noFolderFoundText)}</div>}
-                    </div>
-                    <DialogFooter>
-                        <PrimaryButton onClick={onSelectFolder} text={t(ResourceKeys.modelRepository.types.local.folderPicker.command.select)} disabled={!currentFolder}/>
-                        <DefaultButton onClick={dismissFolderPicker} text={t(ResourceKeys.modelRepository.types.local.folderPicker.command.cancel)} />
-                    </DialogFooter>
-                </Dialog>
-            </div>
-        );
     };
 
     return (
