+++ src/app/shared/modelRepository/interface.ts
@@ -2,7 +2,7 @@
  * Copyright (c) Microsoft Corporation. All rights reserved.
  * Licensed under the MIT License
  **********************************************************/
-import { ModelRepositoryStateInterface } from "./state";
+import { ModelRepositoryStateInterface } from './state';
 
 export interface ModelRepositoryInterface {
     setRepositoryLocations(settings: ModelRepositoryStateInterface): void;
